# apps/app/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SkyPlanner API
  description: |
    REST API for SkyPlanner kundeadministrasjon og ruteplanlegging.

    ## Autentisering

    API-et stotter to autentiseringsmetoder:

    ### API-nokkel
    For tredjeparts-integrasjoner. Opprett API-nokkel i admin-panelet.
    ```
    X-API-Key: sk_live_xxx
    ```

    ### JWT Bearer Token
    For innloggede brukere via web-applikasjonen.
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - Standard: 100 requests per 15 minutter
    - API-nokler kan ha egne kvote-innstillinger

    ## Feilhandtering
    Alle feil returnerer et standardisert format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Beskrivende feilmelding"
      },
      "requestId": "abc123"
    }
    ```

  version: 1.0.0
  contact:
    name: Efffekt AS
    email: support@skyplanner.no
    url: https://skyplanner.no
servers:
  - url: https://app.skyplanner.no/api
    description: Produksjonsserver
  - url: http://localhost:3000/api
    description: Lokal utvikling
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /webhooks:
    get:
      tags:
        - Webhooks
      summary: List webhooks
      description: Henter alle webhook-endepunkter for organisasjonen
      operationId: listWebhooks
      security:
        - ApiKeyAuth: [webhooks:manage]
        - BearerAuth: []
      responses:
        '200':
          description: Liste over webhooks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/WebhookEndpoint'

    post:
      tags:
        - Webhooks
      summary: Opprett webhook
      description: |
        Oppretter et nytt webhook-endepunkt.

        **Viktig:** Secret-verdien returneres kun ved opprettelse og må lagres trygt.
      operationId: createWebhook
      security:
        - ApiKeyAuth: [webhooks:manage]
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook opprettet
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          webhook:
                            $ref: '#/components/schemas/WebhookEndpoint'
                          secret:
                            type: string
                            description: HMAC secret for signaturverifisering (vises kun en gang)
                            example: whsec_abc123...
                      message:
                        type: string
  /webhooks/{id}:
    get:
      tags:
        - Webhooks
      summary: Hent webhook
      operationId: getWebhook
      security:
        - ApiKeyAuth: [webhooks:manage]
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Webhook-detaljer
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WebhookEndpoint'

    put:
      tags:
        - Webhooks
      summary: Oppdater webhook
      operationId: updateWebhook
      security:
        - ApiKeyAuth: [webhooks:manage]
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Webhook oppdatert

    delete:
      tags:
        - Webhooks
      summary: Slett webhook
      operationId: deleteWebhook
      security:
        - ApiKeyAuth: [webhooks:manage]
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Webhook slettet
...
  /webhooks/events:
    get:
      tags:
        - Webhooks
      summary: Tilgjengelige events
      description: Henter liste over tilgjengelige webhook event-typer
      operationId: getWebhookEvents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste over event-typer
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            event:
                              type: string
                              example: customer.created
                            label:
                              type: string
                              example: Kunde opprettet

  # ============ API Keys ============
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API-nokkel opprettet i admin-panelet

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-token fra innlogging
  parameters:
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Unik ID
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        requestId:
          type: string
          example: req_abc123
...
    WebhookEndpoint:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
          format: uri
        name:
          type: string
        description:
          type: string
        events:
          type: array
          items:
            type: string
            enum:
              - customer.created
              - customer.updated
              - customer.deleted
              - route.created
              - route.completed
              - appointment.created
              - appointment.completed
              - sync.completed
              - sync.failed
        is_active:
          type: boolean
        failure_count:
          type: integer
        last_success_at:
          type: string
          format: date-time
        last_failure_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateWebhookRequest:
      type: object
      required:
        - url
        - name
        - events
      properties:
        url:
          type: string
          format: uri
          description: HTTPS URL for webhook-leveranser
        name:
          type: string
          description: Beskrivende navn
        description:
          type: string
        events:
          type: array
          items:
            type: string
          minItems: 1
          description: Event-typer å abonnere på
    UpdateWebhookRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        name:
          type: string
        description:
          type: string
        events:
          type: array
          items:
            type: string
        is_active:
          type: boolean
...